+++ NEWS.md
@@ -1,4 +1,4 @@
-# phonetisr v0.0.3.9000
+# phonetisr v0.0.4
 
 ## Added